+++ src/logging/index.ts
@@ -1,0 +1,2 @@
+export { logger } from './logging-request-scope';
+export { rootLogger, createLogger, BunyanLogger } from './logging';